+++ src/main/java/com/example/randomdriveproject/config/WebSecurityConfig.java
@@ -66,12 +66,6 @@ public class WebSecurityConfig {
         http.authorizeHttpRequests((authorizeHttpRequests) ->
                 authorizeHttpRequests
                         .requestMatchers(PathRequest.toStaticResources().atCommonLocations()).permitAll() // resources 접근 허용 설정
-//                        .requestMatchers("/").permitAll() // 메인 페이지 요청 허가
-//                        .requestMatchers("/navigation.js").permitAll() // 메인 페이지 요청 허가
-//                        .requestMatchers("/navigation").permitAll() // 메인 페이지 요청 허가
-//                        .requestMatchers("/api/user/**").permitAll() // '/api/user/'로 시작하는 요청 모두 접근 허가
-//                        .anyRequest().authenticated() // 그 외 모든 요청 인증처리
-//                        .requestMatchers("/route/**").authenticated()
                         .requestMatchers("/").authenticated()
                         .anyRequest().permitAll()
