+++ src/main/java/com/example/randomdriveproject/controller/RandomRouteController.java
@@ -1,19 +1,16 @@
 package com.example.randomdriveproject.controller;
 
 
-import com.example.randomdriveproject.request.dto.KakaoRouteAllResponseDto;
 import com.example.randomdriveproject.navigation.random.service.RandomKakaoRouteSearchService;
+import com.example.randomdriveproject.request.dto.KakaoRouteAllResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Tag(name = "Random Route Controller", description = "랜덤 경로 네비게이션")
