+++ src/main/java/com/example/randomdriveproject/controller/RouteController.java
@@ -2,15 +2,11 @@ package com.example.randomdriveproject.controller;
 
 import com.example.randomdriveproject.navigation.basic.service.KakaoRouteSearchService;
 import com.example.randomdriveproject.request.dto.KakaoRouteAllResponseDto;
-import com.example.randomdriveproject.user.security.UserDetailsImpl;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
@@ -21,15 +17,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class RouteController {
 
     private final KakaoRouteSearchService kakaoRouteSearchService;
-
-    @Value("${kakao.javascript.api.key}")
-    private String kakaoJavascriptApiKey;
-
-    @GetMapping("/map")
-    public String showMap(Model model) {
-        model.addAttribute("kakaoJavascriptApiKey", kakaoJavascriptApiKey);
-        return "navigation";
-    }
 
     @GetMapping("/route")
     public ResponseEntity<KakaoRouteAllResponseDto> getRoute(@RequestParam String originAddress,
