--- src/main/java/com/example/randomdriveproject/exception/RestApiException.java
@@ -1,11 +1,0 @@
-//package com.example.randomdriveproject.exception;
-//
-//import lombok.AllArgsConstructor;
-//import lombok.Getter;
-//
-//@Getter
-//@AllArgsConstructor
-//public class RestApiException {
-//    private String errorMessage;
-//    private int statusCode;
-//}
