+++ src/main/resources/static/js/histories.js
