+++ src/main/resources/templates/history/histories.html
@@ -27,7 +27,7 @@
     <meta name="viewport" content="initial-scale=1.0, user-scalable=no" width="device-width">
     <script src="https://use.fontawesome.com/798759b2a2.js"></script>
 
-    <link rel="stylesheet" href="/histories.css">
+    <link rel="stylesheet" href="/css/histories.css">
 
     <title>All Histories</title>
 </head>
@@ -62,7 +62,7 @@
         integrity="sha256-2Pmvv0kuTBOenSvLm6bvfBSSHrUJ+3A7x6P5Ebd07/g=" crossorigin="anonymous"></script>
 <script src="https://cdn.jsdelivr.net/npm/js-cookie@3.0.5/dist/js.cookie.min.js"></script>
 
-<script src="/histories.js"></script>
+<script src="/js/histories.js"></script>
 
 </body>
 </html>
