+++ src/main/resources/templates/navi.html
@@ -31,7 +31,7 @@
   <meta name="viewport" content="initial-scale=1.0, user-scalable=no" width="device-width">
   <script src="https://use.fontawesome.com/798759b2a2.js"></script>
 
-  <link rel="stylesheet" href="/navi.css">
+  <link rel="stylesheet" href="/css/navi.css">
 </head>
 <body>
 
@@ -81,7 +81,7 @@
         src="//dapi.kakao.com/v2/maps/sdk.js?appkey=c87b1c2e1d4bcdc3401e35160e8e7d95&libraries=services,clusterer,drawing"></script>
 
 
-<script src="/navi.js"></script> <!-- 여기에 실제 자바스크립트 파일 경로를 입력하세요 -->
+<script src="/js/navi.js"></script> <!-- 여기에 실제 자바스크립트 파일 경로를 입력하세요 -->
 
 </body>
 </html>
