+++ src/main/resources/templates/randomwaypoints.html
@@ -45,7 +45,7 @@
         </div>
     </div>
 </div>
-<script src="/js/navigation.js"></script> <!-- 여기에 실제 자바스크립트 파일 경로를 입력하세요 -->
+<!-- 여기에 실제 자바스크립트 파일 경로를 입력하세요 -->
 <script src="//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js"></script>
 <!-- 카카오 우편번호 서비스 -->
